+++ projects/sip-alain/CHANGELOG.md
@@ -5,6 +5,7 @@
 
 * 优化：sip-cache.service
 * 优化：ng生命周期事件
+* 添加: SipPrepareData && SipInit
 
 ## 0.6.0